converted raw exif button to listTile
